+++ C++/Yandex/src/2_red_belt/week_2/week_2.h
@@ -35,7 +35,7 @@ namespace week_2
 					++uniq_count;
 				}
 
-			return uniq_count;
+			return static_cast<int>(uniq_count);
 		}
 
 		vector<string> KnownWords() {
